# openapi.yaml
openapi: 3.0.3
info:
  title: WUNCoin Blockchain API
  version: 1.0.0
  description: |
    REST + WebSocket API over the WUNCoin blockchain: Proof-of-Work consensus,
    ECDSA-signed transactions, smart contracts (WUNCoin / Staking; MultiSig is
    planned and not yet implemented),
    optional Postgres persistence and an optional Redis cross-replica scaling
    layer.

    ## Conventions
    - Every JSON route returns a `{ success, ... }` envelope. Handled errors use
      `{ "success": false, "error": "<message>" }` with an appropriate status.
    - **Authentication**: protected routes require the `x-api-key` header. The
      gate FAILS CLOSED in production (503 when `BLOCKCHAIN_API_KEY` is unset,
      401 on a wrong/missing key). In dev/test with no key configured the gate is
      permissive.
    - **Rate limiting**: most routes are behind a per-IP limiter (default
      120 req / 60 s, `429` on breach). It is bypassed when `NODE_ENV=test`.
      Redis-backed (shared across replicas) when `REDIS_URL` is set.
    - **Read caching**: read-only blockchain endpoints set
      `Cache-Control: public, max-age=5`.
    - **Request bodies** are capped at `100kb`.
    - A future `/api/v1/...` prefix is reserved; existing paths are unchanged.

    ## WebSockets
    In addition to REST, the service exposes a WebSocket hub on the same port for
    `transaction_added` and `block_mined` events. When `REQUIRE_WS_AUTH=true`,
...

    ## Optional subsystems (dormant by default)
...
  contact:
...
servers:
...
security: []
paths:
...
  /api/blockchain/info:
    get:
      tags: [blockchain]
      summary: Blockchain information
      operationId: getBlockchainInfo
      responses:
        "200":
          description: Chain summary.
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/BlockchainInfo"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /api/blockchain/chain:
    get:
      tags: [blockchain]
      summary: Full blockchain
      operationId: getChain
      responses:
        "200":
          description: The full chain.
          content:
            application/json:
              schema:
                type: object
                required: [success, data, length]
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Block"
                  length:
                    type: integer
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /api/transactions/sign:
    post:
      tags: [transactions]
      summary: Sign a transaction (server-side)
      description: >
        Convenience signing endpoint. DISABLED in production unless
        `ALLOW_SIGN_ENDPOINT=true` (returns 403). Clients should normally sign
        locally and submit to `POST /api/transactions`.
      operationId: signTransaction
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [transaction, privateKey]
              properties:
                transaction:
                  $ref: "#/components/schemas/TransactionInput"
                privateKey:
                  type: string
                  description: Hex private key used to sign.
      responses:
        "200":
          description: Signature produced.
          content:
            application/json:
              schema:
                type: object
                required: [success, signature, message]
                properties:
                  success:
                    type: boolean
                  signature:
                    type: string
                  message:
                    type: string
        "400":
          description: Missing transaction or privateKey.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Signing endpoint disabled in production.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
        "503":
          $ref: "#/components/responses/KeyNotConfigured"
...
  /api/transactions/pending:
    get:
      tags: [transactions]
      summary: Pending transaction count
      description: O(1) count of the mempool (avoids a full chain validation).
      operationId: getPendingCount
      responses:
        "200":
          description: Pending count.
          content:
            application/json:
              schema:
                type: object
                required: [success, pendingCount, message]
                properties:
                  success:
                    type: boolean
                  pendingCount:
                    type: integer
                  message:
                    type: string
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /api/balance/{address}:
    get:
      tags: [balance]
      summary: Address balance and nonce
      operationId: getBalance
      parameters:
        - $ref: "#/components/parameters/Address"
      responses:
        "200":
          description: Balance and contract-state nonce.
          content:
            application/json:
              schema:
                type: object
                required: [success, address, balance, nonce]
                properties:
                  success:
                    type: boolean
                  address:
                    type: string
                  balance:
                    type: number
                  nonce:
                    type: integer
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /api/logs/export:
    post:
      tags: [logs]
      summary: Export logs as a downloadable JSON file
      operationId: exportLogs
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: >
            The exported log array as an attachment
            (`Content-Disposition: attachment; filename="blockchain-logs-<ts>.json"`).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LogEntry"
          headers:
            Content-Disposition:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
        "503":
          $ref: "#/components/responses/KeyNotConfigured"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: >
        Shared secret from `BLOCKCHAIN_API_KEY`. Required in production (fails
        closed with 503 when unset); permissive in dev/test when unconfigured.
  parameters:
    Address:
      name: address
      in: path
      required: true
      description: >
        Account address. ECDSA addresses are `0x` + 40 hex chars; well-known
        accounts such as `TREASURY` are also accepted.
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Missing or invalid API key.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    KeyNotConfigured:
      description: Production API key not configured (fail-closed).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Rate limit exceeded.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error: Rate limit exceeded
    ServerError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string
...
    BlockchainInfo:
      type: object
      required: [chainLength, pendingTransactions, difficulty, validators, consensus, isValid]
      properties:
        chainLength:
          type: integer
        pendingTransactions:
          type: integer
        difficulty:
          type: integer
        validators:
          type: array
          items:
            type: string
        consensus:
          type: string
          example: POW
        isValid:
          type: boolean
        writeDegraded:
          type: boolean
    TransactionType:
      type: string
      enum: [transfer, mint, burn, stake, anchor]
    TransactionInput:
      type: object
      description: >
        Inbound transaction payload. `from`, `to`, `amount`, `type` and `nonce`
        are required. Signed (non-mint, non-TREASURY) transfers must also carry
        `id`, `timestamp`, `signature` and `publicKey`.
      required: [from, to, amount, type, nonce]
      properties:
        id:
          type: string
        from:
          type: string
        to:
          type: string
        amount:
          oneOf:
            - type: number
            - type: string
          description: Numeric (parsed with parseFloat).
        nonce:
          oneOf:
            - type: integer
            - type: string
        timestamp:
          oneOf:
            - type: integer
            - type: string
        type:
          $ref: "#/components/schemas/TransactionType"
        publicKey:
          type: string
        signature:
          type: string
        metadata:
          type: object
          maxProperties: 16
          additionalProperties: true
          description: >
            Optional free-form annotation. It is stored permanently, hashed into
            the block `txRoot` and covered by the sender signature, so the server
            validates it and never normalizes it (any rewrite would desynchronize
            the signed bytes). Bounds: flat object of primitive values only, at
            most 16 keys, snake_case keys up to 32 chars, string values up to 256
            chars, `JSON.stringify(metadata)` up to 2048 bytes. Values JSON would
            drop or rewrite (undefined, functions, NaN, Infinity) are rejected.
            `anchor` transactions additionally require a positive integer
            `anchor_version`. Violations return 400.
    Transaction:
      type: object
      required: [id, from, to, amount, timestamp, nonce, type]
      properties:
        id:
          type: string
        from:
          type: string
        to:
          type: string
        amount:
          type: number
        timestamp:
          type: integer
          format: int64
        nonce:
          type: integer
        publicKey:
          type: string
        signature:
          type: string
        type:
          $ref: "#/components/schemas/TransactionType"
        metadata:
          type: object
          additionalProperties: true
          description: >
            Stored annotation; always satisfies the inbound metadata policy
            documented on TransactionInput.
...
    Block:
      type: object
      required: [index, timestamp, transactions, previousHash, hash, nonce, miner, difficulty]
      properties:
        index:
          type: integer
        timestamp:
          type: integer
          format: int64
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        previousHash:
          type: string
        hash:
          type: string
        nonce:
          type: integer
        miner:
          type: string
        difficulty:
          type: integer
        version:
          type: integer
          description: 1 = legacy full-JSON hashing, 2 = txRoot-based hashing.
        txRoot:
          type: string
          description: Transaction root (version 2 blocks only).
        metadata:
          type: object
          additionalProperties: true
          description: Genesis carries `{ chainId }`.
...
    LogEntry:
      type: object
      properties:
        level:
          type: string
        message:
          type: string
        timestamp:
          type: string
        details:
          type: object
          additionalProperties: true
...